Dodano czesc zadania 13.12 - jeszcze zle obliczane jest dx i dy

+++ PPP2/PPP2GUI/firstExample/RegularPolygon/RegularPolygon.h
@@ -53,7 +53,7 @@ namespace Graph_lib
         int nSides() const { return nSides_; }
 
         //- Returns internal angle
-        int internalAngle() const { return angle_; } 
+        double internalAngle() const { return angle_; } 
 
         //- Returns length of a side
         int side() const { return side_; }
